add thermostat sample changes
