+++ iothub/service/samples/PnpServiceSamples/TemperatureController/TemperatureController.csproj
@@ -6,7 +6,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Microsoft.Azure.Devices.csproj" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="3.1.5" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\src\Microsoft.Azure.Devices.csproj" />
   </ItemGroup>
 
 </Project>
