+++ iothub/service/samples/PnpServiceSamples/Thermostat/Program.cs
@@ -2,10 +2,9 @@
 using System.Threading.Tasks;
 using Microsoft.Azure.Devices;
 using Microsoft.Azure.Devices.Shared;
-using Newtonsoft.Json;
-using PnpHelpers;
+using Microsoft.Extensions.Logging;
 
-namespace TemperatureController
+namespace Thermostat
 {
     public class Program
     {
@@ -13,22 +12,30 @@ namespace TemperatureController
         private static readonly string s_hubConnectionString = Environment.GetEnvironmentVariable("IOTHUB_CONNECTION_STRING");
         private static readonly string s_deviceId = Environment.GetEnvironmentVariable("DEVICE_ID");
 
-        // These are values as defined in DTMI used for device client sample with component.
-        private const string ComponentName = "thermostat1";
-
-        // Writable property to update
+        // These are values as defined in DTMI used for PnP no Component device client sample.
+        // DTDL interface used: https://github.com/Azure/opendigitaltwins-dtdl/blob/master/DTDL/v2/samples/Thermostat.json
         private const string PropertyName = "targetTemperature";
         private const double PropertyValue = 60;
-
-        // Method on a given component
-        private const string MethodName = "getMaxMinReport";
-        private static readonly DateTime dateTime = DateTime.Now ;
+        private const string MethodToInvoke = "reboot";
+        private const string MethodPayload = "{\"delay\":10}";
 
         private static ServiceClient s_serviceClient;
         private static RegistryManager s_registryManager;
+        private static ILogger s_logger;
 
-        public static async Task Main(string[] args)
+        public static async Task Main(string[] _)
         {
+            ILoggerFactory loggerFactory = LoggerFactory.Create(builder =>
+            {
+                builder
+                .AddFilter(level => level >= LogLevel.Debug)
+                .AddConsole(options =>
+                {
+                    options.TimestampFormat = "[MM/dd/yyyy HH:mm:ss]";
+                });
+            });
+            s_logger = loggerFactory.CreateLogger<Program>();
+
             await RunSampleAsync();
         }
 
@@ -37,28 +44,22 @@ namespace TemperatureController
             PrintLog($"Initialize the service client.");
             InitializeServiceClient();
 
-            PrintLog($"Get model Id and Update Component Property");
+            PrintLog($"Get Twin model Id and Update Twin");
             await GetAndUpdateTwinAsync();
 
-            PrintLog($"Invoke a method on a Component");
+            PrintLog($"Invoke a method");
             await InvokeMethodAsync();
         }
 
         private static async Task InvokeMethodAsync()
         {
-            // Create method name to invoke for component
-            string methodToInvoke = PnpHelper.CreatePnpCommandName(MethodName, ComponentName);
-            var methodInvocation = new CloudToDeviceMethod(methodToInvoke) { ResponseTimeout = TimeSpan.FromSeconds(30) };
-
-            // Set Method Payload    
-            var componentMethodPayload = PnpHelper.CreatePnpCommandRequestPayload(JsonConvert.SerializeObject(dateTime));
-            methodInvocation.SetPayloadJson(componentMethodPayload);
-
+            var methodInvocation = new CloudToDeviceMethod(MethodToInvoke) { ResponseTimeout = TimeSpan.FromSeconds(30) };
+            methodInvocation.SetPayloadJson(MethodPayload);
             CloudToDeviceMethodResult result = await s_serviceClient.InvokeDeviceMethodAsync(s_deviceId, methodInvocation);
 
-            if (result == null)
+            if(result == null)
             {
-                throw new Exception("Method invoke returns null");
+                throw new Exception($"Method {MethodToInvoke} invovation returned null");
             }
 
             PrintLog("Method result status is: " + result.Status);
@@ -71,9 +72,8 @@ namespace TemperatureController
             Console.WriteLine("Model Id of this Twin is: " + twin.ModelId);
 
             // Update the twin
-            string propertyUpdate = PnpHelper.CreatePropertyPatch(PropertyName, JsonConvert.SerializeObject(PropertyValue), ComponentName);
-            string twinPatch = $"{{ \"properties\": {{\"desired\": {propertyUpdate}}}}}";
-
+            Twin twinPatch = new Twin();
+            twinPatch.Properties.Desired[PropertyName] = PropertyValue;
             await s_registryManager.UpdateTwinAsync(s_deviceId, twinPatch, twin.ETag);
         }
